[refactor] 웨이팅 다이얼로그 부스 상세화면에서 분리
